pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diagnoai"
version = "0.1.0"
description = "A medical image classification application using AI"
readme = "README.md"
requires-python = ">=3.10,<3.13"


authors = [
    {name = "BellBlaze Technologies", email = "contact@bellblaze.com"}
]

dependencies = [
    "absl-py==2.3.1",
    "altair==5.5.0",
    "asttokens==3.0.0",
    "astunparse==1.6.3",
    "attrs==25.3.0",
    "backcall==0.2.0",
    "beautifulsoup4==4.13.5",
    "bleach==6.2.0",
    "blinker==1.9.0",
    "boto3==1.40.36",
    "botocore==1.40.36",
    "cachetools==6.2.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.3.0",
    "colorama==0.4.6",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "docopt==0.6.2",
    "executing==2.2.1",
    "fastjsonschema==2.21.2",
    "flatbuffers==25.2.10",
    "gast==0.6.0",
    "gitdb==4.0.12",
    "GitPython==3.1.45",
    "google-pasta==0.2.0",
    "grpcio==1.75.0",
    "h5py==3.14.0",
    "idna==3.10",
    "ipython==8.12.3",
    "jedi==0.19.2",
    "Jinja2==3.1.6",
    "jmespath==1.0.1",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter_client==8.6.3",
    "jupyter_core==5.8.1",
    "jupyterlab_pygments==0.3.0",
    "keras==3.11.3",
    "libclang==18.1.1",
    "Markdown==3.9",
    "markdown-it-py==4.0.0",
    "MarkupSafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mistune==3.1.4",
    "ml_dtypes==0.5.3",
    "namex==0.1.0",
    "narwhals==2.5.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "numpy==2.2.6",
    "opt_einsum==3.4.0",
    "optree==0.17.0",
    "packaging==25.0",
    "pandas==2.3.2",
    "pandocfilters==1.5.1",
    "parso==0.8.5",
    "pickleshare==0.7.5",
    "pillow==11.3.0",
    "pipreqs==0.5.0",
    "platformdirs==4.4.0",
    "prompt_toolkit==3.0.52",
    "protobuf==6.32.1",
    "psycopg2-binary==2.9.10",
    "pure_eval==0.2.3",
    "pyarrow==21.0.0",
    "pydeck==0.9.1",
    "pydicom==3.0.1",
    "Pygments==2.19.2",
    "PyJWT==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyzmq==27.1.0",
    "referencing==0.36.2",
    "requests==2.32.5",
    "rich==14.1.0",
    "rpds-py==0.27.1",
    "s3transfer==0.14.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "soupsieve==2.8",
    "stack-data==0.6.3",
    "streamlit==1.49.1",
    "tenacity==9.1.2",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.20.0",
    "termcolor==3.1.0",
    "tinycss2==1.4.0",
    "toml==0.10.2",
    "tornado==6.5.2",
    "traitlets==5.14.3",
    "typing_extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3>=1.25.4,<1.27",
    "watchdog==6.0.0",
    "wcwidth==0.2.14",
    "webencodings==0.5.1",
    "Werkzeug==3.1.3",
    "wrapt==1.17.3",
    "yarg==0.1.9"
]

[project.optional-dependencies]
dev = [
    "ipython==8.12.3",
    "jupyter_client==8.6.3",
    "pipreqs==0.5.0"
]

[project.scripts]
diagnoai = "app:main"

[tool.setuptools]
packages = ["app", "auth_utils", "db_utils", "ui_utils", "aws_secrets_utils"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]


